requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
cloudpickle==2.0.0
dask==2021.11.2
fsspec==2021.11.0
google-api-core==2.2.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.29.0
google-cloud-bigquery-storage==2.9.1
google-cloud-core==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
libcst==0.3.21
locket==0.2.1
mypy-extensions==0.4.3
numpy==1.21.3
oauthlib==3.1.1
packaging==21.2
pandas==1.3.4
pandas-gbq==0.15.0
partd==1.2.0
proto-plus==1.19.7
protobuf==3.19.1
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.6
six==1.16.0
toolz==0.11.2
tqdm==4.62.3
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
